Moved pyqt5 integration into library, cleaning up examples
