+++ radiant/inputs.py
@@ -1,11 +1,6 @@
 from collections import defaultdict
 
 
-__all__ = ('mouse_wheel_delta', 'mouse_position', 'input_string',
-           'key_down', 'key_held', 'key_up',
-           'mouse_button_down', 'mouse_button_held', 'mouse_button_up',
-           'reset')
-
 # window managers should update these values
 mouse_wheel_delta = [0., 0.]
 mouse_position = (0., 0.)
@@ -20,7 +15,7 @@ mouse_button_held = defaultdict(lambda: False)  # True while a button is held do
 mouse_button_up = defaultdict(lambda: False)  # True if it was released in the current event loop iteration
 
 
-def reset():
+def tick():
     """Resets the per-frame values."""
     global mouse_wheel_delta, mouse_position, input_string
     mouse_wheel_delta[0] = 0.
