--- radiant/renderers/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-"""All rendering backends."""
-# not imported automatically, because the dependencies are optional
